x86_64: fix LP64 format/cast errors in drivers and servers, add amd64 distribution sets
Fix LP64 type errors across storage drivers, TTY drivers, and servers needed
to build MINIX for x86_64:

Drivers:
- ahci: %llu→%lu, %llx→%lx, uint32_t→void* cast via uintptr_t
- fbd/vnd/memory: ssize_t return type for bdr_transfer; cascade through
  rule_post_hook/action_post_hook chain; %d→%zu/%zd for size_t/ssize_t
- filter/sum/util: %d→%zu for size_t
- virtio_blk: %llx→%lx, %llu→%lu
- at_wini: %d→%td for ptrdiff_t (wn - wini)
- memory: objcopy -Bi386 for x86_64 targets (BFD arch name)
- net drivers (3c90x, atl2, dp8390, e1000, fxp, lance): LP64 format/cast fixes
- ti1225, pty: int-to-pointer cast via uintptr_t

TTY / console (x86_64):
- Delete bios_console.c; remove BIOS console branch and extern from console.c
- Add beep_x/toggle_scroll no-op stubs to console.c
- Fix paddr_t availability: #define _KERNTYPES in genfb.c, fb_console.c,
  wsdisplay_vcons.c; fallback typedef in wsdisplayvar.h
- cp_grant_id_t → (char *)(uintptr_t) in rs232.c, ser_console.c, fb_console.c

Servers/lib:
- vfs, ipc, pm, vm: LP64 format/sign-comparison/cast fixes
- liblwip: lwipopts.h amd64 type widths
- trace/kernel.c: format spec fixes

Build:
- share/mk/bsd.prog.mk: add libacpi to MINIX libs for-loop
- minix/lib/Makefile: gate libacpi and libvirtio on i386 || x86_64
- Makefile (root): do-lib runs build_install on minix/lib; checkflist uses
  SLOPPY_FLIST=YES (MINIX sets don't enumerate all NetBSD files in DESTDIR)
- etc/mtree/NetBSD.dist.base: add ./usr/libdata/ldscripts directory

Distribution sets (new for amd64):
- distrib/sets/lists/minix-base/md.amd64: services, system.conf.d, libacpi,
  libvirtio, bootx64.efi
- distrib/sets/lists/minix-comp/md.amd64: usr/include/amd64/, usr/include/x86/,
  clang-3.6 intrinsics, ldscripts
- distrib/sets/lists/minix-debug/md.amd64: debug libs and service debug files
- distrib/sets/mkvars.mk: add MKBINUTILS, MKLLVM, MKLLVMCMDS, MKNLS to
  MKEXTRAVARS so tagged files are properly included in the flist
- distrib/sets/lists/minix-base/mi, minix-man/mi: remove netstat (not built)

+++ minix/drivers/storage/memory/memory.c
@@ -414,7 +414,7 @@ static struct device *m_block_part(devminor_t minor)
 /*===========================================================================*
  *				m_block_transfer			     *
  *===========================================================================*/
-static int m_block_transfer(
+static ssize_t m_block_transfer(
   devminor_t minor,		/* minor device number */
   int do_write,			/* read or write? */
   u64_t position,		/* offset on device to read or write */
@@ -569,7 +569,7 @@ static int m_block_ioctl(devminor_t minor, unsigned long request,
 		size -= l;
 	}
 	size = rounddown(size, PAGE_SIZE);
-	r = munmap((void *) a, size);
+	r = munmap((void *)(uintptr_t) a, size);
 	if(r != OK) {
 		printf("memory: WARNING: munmap failed: %d\n", r);
 	}
